requirements.txt
beautifulsoup4
requests
langchain
openai
pydantic
python-dotenv
lxml




